dashboard, components, and queries

+++ native/src/screens/dashboard/components/category-item/category-item.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Pressable } from 'react-native'
-import { Symbol, IconButton } from '../../../../components'
+import { Symbol, Icon } from '../../../../components'
 import { Dashboard_items } from '../../types/Dashboard'
 import { useDeleteItem } from './use-delete-item'
 import { useUpdateItem } from './use-update-item'
@@ -11,20 +11,27 @@ import {
 } from './style'
 
 interface Props {
+  isDeleteVisible: boolean
   item: Dashboard_items
+  showDelete: () => void
 }
 
-export const CategoryItem: React.FC<Props> = ({ item }) => {
+export const CategoryItem: React.FC<Props> = ({
+  isDeleteVisible,
+  item,
+  showDelete,
+}) => {
   const { deleteItem } = useDeleteItem()
   const { updateItem } = useUpdateItem()
   const { id, name, purchased } = item
 
-  const handleUpdate = async () => {
-    await updateItem(id)
-  }
+  const handlePress = async () => {
+    if (isDeleteVisible) {
+      await deleteItem(id)
+      return
+    }
 
-  const handleDelete = async () => {
-    await deleteItem(id)
+    await updateItem(id)
   }
 
   return (
@@ -32,18 +39,18 @@ export const CategoryItem: React.FC<Props> = ({ item }) => {
       <Pressable
         accessibilityHint="Update Item"
         accessibilityLabel={purchased ? 'Purchased' : 'Not Purchased'}
-        onPress={handleUpdate}
+        onLongPress={showDelete}
+        onPress={handlePress}
       >
         <CategoryItemCheckbox>
-          <Symbol symbol={purchased ? 'checkmark' : 'unselected'} />
-          <CategoryItemText strikethrough={purchased}>{name}</CategoryItemText>
+          {!isDeleteVisible ? (
+            <Symbol symbol={purchased ? 'checkmark' : 'unselected'} />
+          ) : (
+            <Icon color="light-grey" height={20} icon="trash" width={20} />
+          )}
+          <CategoryItemText purchased={purchased}>{name}</CategoryItemText>
         </CategoryItemCheckbox>
       </Pressable>
-      <IconButton
-        accessibilityLabel="Delete Item"
-        icon="trash"
-        onPress={handleDelete}
-      />
     </CategoryItemContainer>
   )
 }
